pyproject.toml
[project]
name = "vibepod"
version = "0.1.0"
description = "One CLI for containerized AI coding agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "VibePod" }]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "docker>=7.0.0",
    "pyyaml>=6.0.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
vp = "vibepod.cli:app"
vibepod = "vibepod.cli:app"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibepod"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
